# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DocFlow Document Processing API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DocumentStatus =
  (typeof DocumentStatus)[keyof typeof DocumentStatus];

export const DocumentStatus = {
  queued: "queued",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export type DocumentExtractedData = { [key: string]: unknown } | null;

export interface Document {
  id: string;
  filename: string;
  fileSize: number;
  mimeType: string;
  status: DocumentStatus;
  /**
   * @minimum 0
   * @maximum 100
   */
  progress: number;
  uploadedAt: string;
  processedAt?: string | null;
  extractedData?: DocumentExtractedData;
  errorMessage?: string | null;
}

export type UpdateDocumentRequestExtractedData = { [key: string]: unknown };

export interface UpdateDocumentRequest {
  extractedData?: UpdateDocumentRequestExtractedData;
}

export interface Stats {
  total: number;
  queued: number;
  processing: number;
  completed: number;
  failed: number;
}

export type UploadDocumentBody = {
  file: Blob;
};

export type ExportDocumentParams = {
  format: ExportDocumentFormat;
};

export type ExportDocumentFormat =
  (typeof ExportDocumentFormat)[keyof typeof ExportDocumentFormat];

export const ExportDocumentFormat = {
  json: "json",
  csv: "csv",
} as const;

export type ExportDocument200One = { [key: string]: unknown };