# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DocFlow Document Processing API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: documents
    description: Document operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all documents
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a new document
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "201":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Document not found
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update document extracted data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentRequest"
      responses:
        "200":
          description: Document updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}/retry:
    post:
      operationId: retryDocument
      tags: [documents]
      summary: Retry a failed document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document queued for retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}/export:
    get:
      operationId: exportDocument
      tags: [documents]
      summary: Export document data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: format
          in: query
          required: true
          schema:
            type: string
            enum: [json, csv]
      responses:
        "200":
          description: Exported document data
          content:
            application/json:
              schema:
                type: object
            text/csv:
              schema:
                type: string
  /stats:
    get:
      operationId: getStats
      tags: [documents]
      summary: Get processing statistics
      responses:
        "200":
          description: Processing stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Document:
      type: object
      properties:
        id:
          type: string
        filename:
          type: string
        fileSize:
          type: number
        mimeType:
          type: string
        status:
          type: string
          enum: [queued, processing, completed, failed]
        progress:
          type: number
          minimum: 0
          maximum: 100
        uploadedAt:
          type: string
          format: date-time
        processedAt:
          type: string
          format: date-time
          nullable: true
        extractedData:
          type: object
          nullable: true
          additionalProperties: true
        errorMessage:
          type: string
          nullable: true
      required:
        - id
        - filename
        - fileSize
        - mimeType
        - status
        - progress
        - uploadedAt
    UpdateDocumentRequest:
      type: object
      properties:
        extractedData:
          type: object
          additionalProperties: true
    Stats:
      type: object
      properties:
        total:
          type: number
        queued:
          type: number
        processing:
          type: number
        completed:
          type: number
        failed:
          type: number
      required:
        - total
        - queued
        - processing
        - completed
        - failed