# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DocFlow Document Processing API
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentExtractedData } from "./documentExtractedData";
import type { DocumentStatus } from "./documentStatus";

export interface Document {
  id: string;
  filename: string;
  fileSize: number;
  mimeType: string;
  status: DocumentStatus;
  /**
   * @minimum 0
   * @maximum 100
   */
  progress: number;
  uploadedAt: Date;
  processedAt?: Date | null;
  extractedData?: DocumentExtractedData;
  errorMessage?: string | null;
}