# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DocFlow Document Processing API
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  total: number;
  queued: number;
  processing: number;
  completed: number;
  failed: number;
}