# lib/api-zod/src/generated/types/updateDocumentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DocFlow Document Processing API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateDocumentRequestExtractedData } from "./updateDocumentRequestExtractedData";

export interface UpdateDocumentRequest {
  extractedData?: UpdateDocumentRequestExtractedData;
}